Initial Upload
Using object-oriented, tested code using the language that you feel most
proficient in, please write a "pico y placa" predictor. The inputs
should be a license plate number (the full number, not the last digit),
a date (as a String), and a time, and the program will return whether or
not that car can be on the road.